requirements.txt
# MedGemma Fine-tuning Requirements
# Core ML/DL libraries
torch>=2.0.0
transformers>=4.47.0
datasets>=3.0.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.11.0

# Quantization and optimization
bitsandbytes>=0.44.0

# Computer vision
Pillow>=10.0.0
opencv-python>=4.8.0

# Evaluation and metrics
scikit-learn>=1.3.0
evaluate>=0.4.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Hugging Face Hub
huggingface-hub>=0.20.0

# Configuration management
pyyaml>=6.0
